Fixes #33434 - Unify fact importers
