--- app/services/foreman_salt/fact_importer.rb
@@ -1,15 +1,0 @@
-module ForemanSalt
-  class FactImporter < ::StructuredFactImporter
-    def fact_name_class
-      ForemanSalt::FactName
-    end
-
-    def self.support_background
-      true
-    end
-
-    def self.authorized_smart_proxy_features
-      'Salt'
-    end
-  end
-end
