--- app/services/puppet_fact_importer.rb
@@ -1,9 +1,0 @@
-class PuppetFactImporter < StructuredFactImporter
-  def self.authorized_smart_proxy_features
-    'Puppet'.freeze
-  end
-
-  def fact_name_class
-    PuppetFactName
-  end
-end
