+++ test/unit/katello/rhsm_normalizer_test.rb
@@ -1,7 +1,7 @@
 require 'test_helper'
 
 module Katello
-  class RhsmFactImporterTest < ActiveSupport::TestCase
+  class NormalizerTest < ActiveSupport::TestCase
     def setup
       @separator = RhsmFactName::SEPARATOR
       @facts = {
@@ -12,9 +12,9 @@ module Katello
       @host = Host::Managed.new(name: "my.great.host", managed: false)
     end
 
-    def test_importer_separator_change
-      importer = Katello::RhsmFactImporter.new(@host)
-      new_facts = importer.change_separator(@facts)
+    def test_normalizer_separator_change
+      normalizer = FactImporters::Normalizers::Katello.new(@host, {})
+      new_facts = normalizer.change_separator(@facts)
 
       assert_equal new_facts['interesting::key'], 'value_one'
       assert_equal new_facts['another::interesting::key'], 'value_two'
