+++ test/unit/structured_fact_importer_test.rb
@@ -56,32 +56,32 @@ class StructuredFactImporterTest < ActiveSupport::TestCase
 
   describe 'normalize' do
     test 'has no effect on unstructured facts' do
-      importer = StructuredFactImporter.new(nil, 'a' => 'b')
+      importer = FactImporters::Structured.new(nil, nil, 'a' => 'b')
       assert_equal({'a' => 'b'}, importer.send(:facts))
     end
 
     test 'removes nil fact values' do
-      importer = StructuredFactImporter.new(nil, 'a' => nil)
+      importer = FactImporters::Structured.new(nil, nil, 'a' => nil)
       assert_equal({}, importer.send(:facts))
     end
 
     test 'keeps false fact values' do
-      importer = StructuredFactImporter.new(nil, 'a' => false)
+      importer = FactImporters::Structured.new(nil, nil, 'a' => false)
       assert_equal({'a' => 'false'}, importer.send(:facts))
     end
 
     test 'changes symbol keys to strings' do
-      importer = StructuredFactImporter.new(nil, :a => 'b')
+      importer = FactImporters::Structured.new(nil, nil, :a => 'b')
       assert_equal({'a' => 'b'}, importer.send(:facts))
     end
 
     test 'expands nested hash keys with separators' do
-      importer = StructuredFactImporter.new(nil, 'a' => {'b' => 'c'})
+      importer = FactImporters::Structured.new(nil, nil, 'a' => {'b' => 'c'})
       assert_equal({'a' => nil, 'a::b' => 'c'}, importer.send(:facts))
     end
 
     test 'changes non-string values to strings' do
-      importer = StructuredFactImporter.new(nil, :a => 1)
+      importer = FactImporters::Structured.new(nil, nil, :a => 1)
       assert_equal({'a' => '1'}, importer.send(:facts))
     end
 
@@ -92,7 +92,7 @@ class StructuredFactImporterTest < ActiveSupport::TestCase
 
       facts = data.good_facts.deep_merge(data.ignored_facts)
 
-      importer = StructuredFactImporter.new(nil, facts)
+      importer = FactImporters::Structured.new(nil, nil, facts)
       actual_facts = importer.send(:facts)
 
       assert_equal data.flat_result, actual_facts
@@ -101,40 +101,40 @@ class StructuredFactImporterTest < ActiveSupport::TestCase
     test 'filters out too big subtrees' do
       input = {:a => 1, :b => {}}
       (1..101).each { |i| input[:b][:"c_#{i}"] = :test }
-      importer = StructuredFactImporter.new(nil, input)
+      importer = FactImporters::Structured.new(nil, nil, input)
       assert_equal({"a" => "1", "foreman::dropped_subtree_facts" => "100", "foreman" => nil}, importer.send(:facts))
     end
 
     test 'does not filter out "small" subtrees' do
       input = {:a => 1, :b => {}}
       (1..3).each { |i| input[:b][:"c_#{i}"] = :test }
-      importer = StructuredFactImporter.new(nil, input)
+      importer = FactImporters::Structured.new(nil, nil, input)
       assert_equal({"a" => "1", "b::c_1" => "test", "b::c_2" => "test", "b::c_3" => "test", "b" => nil}, importer.send(:facts))
     end
 
     test 'does not filter out few flat facts prefixed with blockdevice' do
       input = {:a => 1, :blockdevice_sda => 1}
-      importer = StructuredFactImporter.new(nil, input)
+      importer = FactImporters::Structured.new(nil, nil, input)
       assert_equal({"a" => "1", "blockdevice_sda" => "1"}, importer.send(:facts))
     end
 
     test 'filters out flat facts prefixed with blockdevice_' do
       input = {:a => 1}
       (1..101).each { |i| input[:"blockdevice_#{i}"] = :test }
-      importer = StructuredFactImporter.new(nil, input)
+      importer = FactImporters::Structured.new(nil, nil, input)
       assert_equal({"a" => "1", "foreman::dropped_subtree_facts" => "100", "foreman" => nil}, importer.send(:facts))
     end
 
     test 'filters out flat facts prefixed with macaddress_' do
       input = {:a => 1}
       (1..101).each { |i| input[:"macaddress_#{i}"] = :test }
-      importer = StructuredFactImporter.new(nil, input)
+      importer = FactImporters::Structured.new(nil, nil, input)
       assert_equal({"a" => "1", "foreman::dropped_subtree_facts" => "100", "foreman" => nil}, importer.send(:facts))
     end
   end
 
   def import(facts)
-    @importer = StructuredFactImporter.new(host, facts)
+    @importer = FactImporters::Structured.new(host, nil, facts)
     @importer.stubs(:fact_name_class).returns(FactName)
     allow_transactions_for @importer
     @importer.import!
